requirements.txt
discord.py[voice]
colorama
requests
yt-dlp
wikipedia